Gameplay algorithm update
Changed the way in which moves and merges are performed using recursive
functions. Added matrix.js to libs folder. Still in process of writing
the functions for matrix math/linear algebra. Also changed the html form
around and added play/pause buttons. Next on the list will be to finish
up the matrix library and update the html page to look more
professional.